backend/requirements.txt
# Core Django
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0

# Database
psycopg2-binary>=2.9.5

# Authentication
djangorestframework-simplejwt>=5.2.2
python-jose>=3.3.0
python-dotenv>=0.21.0

# API Documentation
drf-yasg>=1.21.4

# Testing
pytest>=7.3.1
pytest-django>=4.5.2
pytest-cov>=4.0.0

# Code Quality
flake8>=6.0.0
black>=23.3.0
isort>=5.12.0

# Caching
django-redis>=5.2.0

# Async
channels>=4.0.0
channels-redis>=4.0.0

# Search
django-elasticsearch-dsl>=7.2.2

# Utils
python-dateutil>=2.8.2
Pillow>=9.5.0
requests>=2.28.2

# Production
gunicorn>=20.1.0
whitenoise>=6.4.0

# Monitoring
sentry-sdk>=1.22.0

# Development
django-debug-toolbar>=4.1.0
django-extensions>=3.2.1

# 
django-filter
